Add styles for home
