+++ tailwind.config.js
@@ -1,11 +1,22 @@
+const defaultTheme = require('tailwindcss/defaultTheme')
+
 module.exports = {
   purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   darkMode: false, // or 'media' or 'class'
   theme: {
-    extend: {},
+    extend: {
+      fontFamily: {
+        'sans': ['Lato', ...defaultTheme.fontFamily.sans],
+      },
+      colors: {
+        'bh-blue': '#4369B2',
+        'bh-lightBlue': '#5788E5',
+        'bh-black': '#1D1D1D',
+      },
+    },
   },
   variants: {
     extend: {},
   },
   plugins: [],
-}
+};
